Refactored TagsCell into its own component #25
